Split issue properties pane module
