+++ ui/src/components/issue-properties/index.ts
@@ -1,0 +1,5 @@
+export { IssueProperties } from "./IssueProperties";
+export { ExternalObjectRows } from "./external-object-rows";
+export { PropertyPicker } from "./property-picker";
+export { PropertyChip, PropertyRow, PropertySection } from "./primitives";
+export { ExpandRelationListButton, RemovableIssueReferencePill } from "./relation-controls";
